requirements.txt
discord.py==2.6.4
ruff==0.14.10
aiosqlite==0.22.1
